+++ R/neighbor_pooling.R
@@ -48,7 +48,8 @@ pool_across_neighbors <- function(x, neighbors) {
 #' @param feature Character scalar naming the gene you'd like to threshold. 
 #' @template param_operator
 #' @param threshold Numeric scalar with the expression threshold separating positive
-#' from negative cells.
+#' from negative cells. Evaluate_rule expects a plain fraction, not CP10K
+#' (while rule has the use_CP10K argument).
 #' 
 #' @description The rule is defined by feature, operator and threshold.
 #'
@@ -61,7 +62,8 @@ pool_across_neighbors <- function(x, neighbors) {
 evaluate_rule <- function(obj,
                           feature,
                           operator,
-                          threshold) {
+                          threshold,
+                          overdispersion=0.01) {
   # This is a separate function for two reasons:
   #    * evaluate_rule uses Poisson, I was considering NB at some point as well
   #    * it's being used by classify AND by plot_last, so separate function.
@@ -80,8 +82,9 @@ evaluate_rule <- function(obj,
   S <- pool_across_neighbors(obj$totalUMI,
                              obj$neighbors)
     
-
-  cdf <- stats::ppois(K, S*threshold)
+  # before cellpypes version 0.1.14:
+  # cdf <- stats::ppois(K, S*threshold)
+  cdf <- stats::pnbinom(K, mu=S*threshold, size=1/overdispersion)
   switch(operator,
          # >= and <= are currently prevented by stopifnot in rule
          ">" = cdf > .99,